backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
openai==1.57.0
python-dotenv==1.0.1

